Implementation of POD Stopper and Deleter unit tests
